refactor(vue): extract chart theme logic into useChartTheme composable
Move theme state (chartTheme, chartSettings, tooltipColors, themeCssVars),
auto theme detection with MediaQueryList, and settings handler into
packages/vue/src/composables/chart/useChartTheme.ts.
Reduces KLineChart.vue <script setup> by ~50 lines.
